feat: add function collecting data of play time
