+++ src/main/java/racingcar/io/OutputView.java
@@ -1,0 +1,7 @@
+package racingcar.io;
+
+public class OutputView {
+    public static void printMessage(IOMessage ioMessage) {
+        System.out.println(ioMessage.getMessage());
+    }
+}
